requirements.txt
# Install with: pip install -r requirements.txt
# For editable install: pip install -e .
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.28.0
httpx>=0.27
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
gTTS>=2.3
python-dotenv>=1.0
